Module map_editor::model

Structs

Model

Enums

ID

Constants

BUILDING_LENGTH

INTERSECTION_RADIUS

Functions

dump_to_osm
Express a RawMap as a .osm file. Why not just save the RawMap? The format may change over time, and even if a RawMap is saved as JSON, manually updating it is annoying. This is used to create synthetic maps that will never go bad -- there will always be a pipeline to import a .osm file, so actually, .osm is a stable-over-time format.

time_to_id